integrate get statistics into get state

+++ service/models/get-state.go
@@ -6,10 +6,10 @@ type GetStateRequest struct {
 type GetStateResponse struct {
 	ErrId      int
 	State      int
-	KeysStored int
-	Reservable int
-	KeysServed int
-	KeysAdded  int
+	KeysStored uint64
+	Reservable uint64
+	KeysServed uint64
+	KeysAdded  uint64
 	KeyId0     []byte
 	KeyId1     []byte
 }
